pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-failsafe"
version = "0.3.0"
description = "FailSafe governance adapter for Microsoft Agent Governance Toolkit"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Kevin Knapp", email = "kevin@mythologiq.com" },
]
keywords = ["agent", "governance", "failsafe", "security", "trust"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
agent-os = ["agent-os-kernel>=1.1.0"]
agent-sre = ["agent-sre>=1.1.0"]
agent-mesh = ["agentmesh-platform>=1.1.0"]
agent-hypervisor = ["agent-hypervisor>=1.1.0"]
full = [
    "agent-os-kernel>=1.1.0",
    "agent-sre>=1.1.0",
    "agentmesh-platform>=1.1.0",
    "agent-hypervisor>=1.1.0",
]
dev = [
    "agent-os-kernel>=1.1.0",
    "agent-sre>=1.1.0",
    "agentmesh-platform>=1.1.0",
    "agent-hypervisor>=1.1.0",
    "pytest>=7.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/MythologIQ/agent-failsafe"
Repository = "https://github.com/MythologIQ/agent-failsafe"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_failsafe"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
